Paleography reliability (reasoning-only retry, raised-ceiling+disabled-
reasoning retry, cache hygiene), Ann's batch (sidebar chevron/spinner,
model-add persistence, run-menu honesty, island precedence), pane
program (C1/C3/C8/C9/C10), Mandate-1 outline endpoint, empty-upload
guards, ingest retention contract fix.

+++ fichero-server/src/fichero_server/api/routes/workflow_execution/comparison.py
@@ -50,7 +50,9 @@ class ComparisonSideResponse(BaseModel):
     workflow_name: str
     status: str
     error: str | None = None
-    duration_ms: int | None = None
+    # float, not int (2026-08-23): run durations arrive as fractional ms
+    # (3018.59…) and the int field 500'd every Compare Runs click.
+    duration_ms: float | None = None
     artifact_count: int = 0
     steps_total: int = 0
     steps_failed: int = 0
